feat: add class gameButtons for render buttons
