+++ rss-puzzle/src/app/pages/mainPage/gameButtons/gameButtons.scss
@@ -1,0 +1,6 @@
+.game-buttons {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 20px;
+}
